refactor: extract manual-model validation to a non-component module
The AddManualModelDialog test needs validateManualModel, but exporting a
non-component symbol from a component file breaks React fast-refresh
(react-refresh/only-export-components). Move the pure parsing/validation
logic to a sibling manual-model.ts and import it back.
